Adds DESTDIR support
DESTDIR is the preferred (autotool supported) way to create a staged
deployment of autotools based projects.  Using prefix for staging
can break things badly.

Rationale: Complex libraries with interdependencies use prefix to determine
where the library is installed, header files are installed, and
in some cases where related configuration files are installed.
So it is a bad idea to use --prefix to change the installation path
when using it with FFI.  It can break the build system.

This patch is functionally compatible with previous versions of
autotools-rs but provides the ability to use destdir rather
prefix to perform local installation (whenever it's needed) for
complex libraries with interdependencies.